requirements.txt
Django==1.9.5
argparse==1.2.1
dj-database-url==0.4.0
django-extensions==1.6.1
django-save-the-change==1.1.0
django-timezone-field==1.3
gunicorn==19.4.5
psycopg2==2.6.1
pytz==2016.3
six==1.10.0
wsgiref==0.1.2
